// src/api/types/orders.type.ts
import { CartProduct } from "./product.type";

export interface OrderType {
    shippingAddress:   ShippingAddress;
    taxPrice:          number;
    shippingPrice:     number;
    totalOrderPrice:   number;
    paymentMethodType: string;
    isPaid:            boolean;
    isDelivered:       boolean;
    _id:               string;
    user:              User;
    cartItems:         CartProduct[];
    createdAt:         string;
    updatedAt:         string;
    id:                number;
    __v:               number;
}
export interface ShippingAddress {
    details: string;
    phone:   string;
    city:    string;
}
export interface User {
    _id:   string;
    name:  string;
    email: string;
    phone: string;
}